pi/pyproject.toml
[project]
name = "walla"
version = "0.1.0"
description = "Wall-A robot application — runs on Raspberry Pi 5"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "pygame>=2.6.0",
]

[project.scripts]
walla = "walla.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
